fix delays

+++ minigotchi/deauth.h
@@ -26,7 +26,7 @@ public:
     static int randomIndex;
 
 private:
-    static bool send(uint8 *buf, int len, bool sys_seq);
+    static bool send(uint8 *buf, uint16_t len, bool sys_seq);
     static bool broadcast(uint8_t* mac);
     static void printMac(uint8_t* mac);
     static String printMacStr(uint8_t* mac);
